feat(node): implement `zlib`
- feat(node): implement `zlib` module
- test: add brotli test script
- fix: bytesequence support for fs
- fix: verification metadata for jpms repo
- chore: force brotli to load on native bin load
- chore: update `graalvm` api pins
- chore: update detekt baseline
- chore: update reachability configs
- chore: update to latest gvm tools plugin

+++ packages/graalvm/src/main/kotlin/elide/runtime/node/childProcess/NodeChildProcess.kt
@@ -694,7 +694,7 @@ public class ChildProcessHandle private constructor(
   @Suppress("TooGenericExceptionCaught")
   @Polyglot override fun kill(signal: String?): Unit = (signal ?: pending.execution.options.killSignal).let {
     try {
-      ChildProcessNative.killWith(pid.toInt(), it)
+      ChildProcessNative.killWith(pid, it)
       termination.signal(it)
     } catch (e: Throwable) {
       throw JsError.valueError("Failed to send signal to process: ${e.message}", e)
